requirements.txt
flask==2.3.3
python-dotenv==1.0.0
google-generativeai==0.3.1
opencv-python==4.8.0.76
python-docx==0.8.11
Werkzeug==2.3.7
Pillow==10.0.0
requests==2.31.0 